Add more endpoints and simplify test
